Resolve conflict from JPMML upgrade

+++ app/oryx-app/src/test/java/com/cloudera/oryx/app/speed/kmeans/KMeansSpeedIT.java
@@ -76,7 +76,7 @@ public final class KMeansSpeedIT extends AbstractSpeedIT {
     assertTrue(model instanceof ClusteringModel);
 
     ClusteringModel clusteringModel = (ClusteringModel) model;
-    assertEquals(NUM_CLUSTERS, clusteringModel.getNumberOfClusters().intValue());
+    assertEquals(NUM_CLUSTERS, clusteringModel.getNumberOfClusters());
     List<Cluster> clusters = clusteringModel.getClusters();
 
     Map<Integer,ClusterInfo> clusterInfos = new HashMap<>();
